+++ Flow/Changes.md
@@ -74,6 +74,7 @@
 - Added **TimedTask** trait, which is now supported in **TimedTasks**, also
 - Added new ways to write **LocalTime** values (after importing **TimeExtensions**)
 - Added **SettableOnce**, which is like a **Promise** with **ChangeEvent**s
+- Added **ReleasingPointer** class
 - Added **Identity** object which functions as an identity function (i.e. `a => a`)
 - Added **NoOpLogger** object
 - Added **ApproxSelfEquals** trait